Refactoring on bean classes
- Standardization of the class names with suffix "Bean"
- Generated automatic serialVersionUID
